+++ tests/DependencyInjection/AiBundleTest.php
@@ -57,6 +57,7 @@ use Symfony\AI\Store\Bridge\ChromaDb\Store as ChromaDbStore;
 use Symfony\AI\Store\Bridge\ChromaDb\StoreFactory as ChromaDbStoreFactory;
 use Symfony\AI\Store\Bridge\ClickHouse\Store as ClickhouseStore;
 use Symfony\AI\Store\Bridge\Cloudflare\Store as CloudflareStore;
+use Symfony\AI\Store\Bridge\Cloudflare\StoreFactory as CloudflareStoreFactory;
 use Symfony\AI\Store\Bridge\ManticoreSearch\Store as ManticoreSearchStore;
 use Symfony\AI\Store\Bridge\MariaDb\Distance as MariaDbDistance;
 use Symfony\AI\Store\Bridge\MariaDb\Store as MariaDbStore;
@@ -1001,15 +1002,16 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasDefinition('ai.store.cloudflare.my_cloudflare_store'));
 
         $definition = $container->getDefinition('ai.store.cloudflare.my_cloudflare_store');
+        $this->assertSame([CloudflareStoreFactory::class, 'create'], $definition->getFactory());
         $this->assertSame(CloudflareStore::class, $definition->getClass());
 
         $this->assertTrue($definition->isLazy());
         $this->assertCount(6, $definition->getArguments());
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(0));
-        $this->assertSame('http_client', (string) $definition->getArgument(0));
+        $this->assertSame('random', $definition->getArgument(0));
         $this->assertSame('foo', $definition->getArgument(1));
         $this->assertSame('bar', $definition->getArgument(2));
-        $this->assertSame('random', $definition->getArgument(3));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('http_client', (string) $definition->getArgument(3));
         $this->assertSame(1536, $definition->getArgument(4));
         $this->assertSame('cosine', $definition->getArgument(5));
 
@@ -1024,10 +1026,8 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasAlias(StoreInterface::class.' $myCloudflareStore'));
         $this->assertTrue($container->hasAlias(StoreInterface::class.' $cloudflareMyCloudflareStore'));
         $this->assertTrue($container->hasAlias(StoreInterface::class));
-    }
 
-    public function testCloudflareStoreWithCustomIndexCanBeConfigured()
-    {
+        // Custom index
         $container = $this->buildContainer([
             'ai' => [
                 'store' => [
@@ -1045,15 +1045,16 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasDefinition('ai.store.cloudflare.my_cloudflare_store'));
 
         $definition = $container->getDefinition('ai.store.cloudflare.my_cloudflare_store');
+        $this->assertSame([CloudflareStoreFactory::class, 'create'], $definition->getFactory());
         $this->assertSame(CloudflareStore::class, $definition->getClass());
 
         $this->assertTrue($definition->isLazy());
         $this->assertCount(6, $definition->getArguments());
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(0));
-        $this->assertSame('http_client', (string) $definition->getArgument(0));
+        $this->assertSame('random', $definition->getArgument(0));
         $this->assertSame('foo', $definition->getArgument(1));
         $this->assertSame('bar', $definition->getArgument(2));
-        $this->assertSame('random', $definition->getArgument(3));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('http_client', (string) $definition->getArgument(3));
         $this->assertSame(1536, $definition->getArgument(4));
         $this->assertSame('cosine', $definition->getArgument(5));
 
@@ -1068,10 +1069,8 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasAlias(StoreInterface::class.' $myCloudflareStore'));
         $this->assertTrue($container->hasAlias(StoreInterface::class.' $cloudflareMyCloudflareStore'));
         $this->assertTrue($container->hasAlias(StoreInterface::class));
-    }
 
-    public function testCloudflareStoreWithCustomDimensionsCanBeConfigured()
-    {
+        // Custom dimensions
         $container = $this->buildContainer([
             'ai' => [
                 'store' => [
@@ -1089,15 +1088,16 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasDefinition('ai.store.cloudflare.my_cloudflare_store'));
 
         $definition = $container->getDefinition('ai.store.cloudflare.my_cloudflare_store');
+        $this->assertSame([CloudflareStoreFactory::class, 'create'], $definition->getFactory());
         $this->assertSame(CloudflareStore::class, $definition->getClass());
 
         $this->assertTrue($definition->isLazy());
         $this->assertCount(6, $definition->getArguments());
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(0));
-        $this->assertSame('http_client', (string) $definition->getArgument(0));
+        $this->assertSame('my_cloudflare_store', $definition->getArgument(0));
         $this->assertSame('foo', $definition->getArgument(1));
         $this->assertSame('bar', $definition->getArgument(2));
-        $this->assertSame('my_cloudflare_store', $definition->getArgument(3));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('http_client', (string) $definition->getArgument(3));
         $this->assertSame(768, $definition->getArgument(4));
         $this->assertSame('cosine', $definition->getArgument(5));
 
@@ -1112,10 +1112,8 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasAlias(StoreInterface::class.' $myCloudflareStore'));
         $this->assertTrue($container->hasAlias(StoreInterface::class.' $cloudflareMyCloudflareStore'));
         $this->assertTrue($container->hasAlias(StoreInterface::class));
-    }
 
-    public function testCloudflareStoreWithCustomEndpointCanBeConfigured()
-    {
+        // Custom endpoint
         $container = $this->buildContainer([
             'ai' => [
                 'store' => [
@@ -1135,18 +1133,58 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasDefinition('ai.store.cloudflare.my_cloudflare_store'));
 
         $definition = $container->getDefinition('ai.store.cloudflare.my_cloudflare_store');
+        $this->assertSame([CloudflareStoreFactory::class, 'create'], $definition->getFactory());
         $this->assertSame(CloudflareStore::class, $definition->getClass());
 
         $this->assertTrue($definition->isLazy());
         $this->assertCount(7, $definition->getArguments());
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(0));
-        $this->assertSame('http_client', (string) $definition->getArgument(0));
+        $this->assertSame('my_cloudflare_store', $definition->getArgument(0));
         $this->assertSame('foo', $definition->getArgument(1));
         $this->assertSame('bar', $definition->getArgument(2));
-        $this->assertSame('my_cloudflare_store', $definition->getArgument(3));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('http_client', (string) $definition->getArgument(3));
         $this->assertSame(1536, $definition->getArgument(4));
         $this->assertSame('cosine', $definition->getArgument(5));
         $this->assertSame('https://api.cloudflare.com/client/v5/accounts', $definition->getArgument(6));
+
+        $this->assertTrue($definition->hasTag('proxy'));
+        $this->assertSame([
+            ['interface' => StoreInterface::class],
+            ['interface' => ManagedStoreInterface::class],
+        ], $definition->getTag('proxy'));
+        $this->assertTrue($definition->hasTag('ai.store'));
+
+        $this->assertTrue($container->hasAlias('.'.StoreInterface::class.' $my_cloudflare_store'));
+        $this->assertTrue($container->hasAlias(StoreInterface::class.' $myCloudflareStore'));
+        $this->assertTrue($container->hasAlias(StoreInterface::class.' $cloudflareMyCloudflareStore'));
+        $this->assertTrue($container->hasAlias(StoreInterface::class));
+
+        // Custom HttpClient
+        $container = $this->buildContainer([
+            'ai' => [
+                'store' => [
+                    'cloudflare' => [
+                        'my_cloudflare_store' => [
+                            'http_client' => 'my.scoped_http_client',
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+
+        $definition = $container->getDefinition('ai.store.cloudflare.my_cloudflare_store');
+        $this->assertSame([CloudflareStoreFactory::class, 'create'], $definition->getFactory());
+        $this->assertSame(CloudflareStore::class, $definition->getClass());
+
+        $this->assertTrue($definition->isLazy());
+        $this->assertCount(6, $definition->getArguments());
+        $this->assertSame('my_cloudflare_store', $definition->getArgument(0));
+        $this->assertNull($definition->getArgument(1));
+        $this->assertNull($definition->getArgument(2));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('my.scoped_http_client', (string) $definition->getArgument(3));
+        $this->assertSame(1536, $definition->getArgument(4));
+        $this->assertSame('cosine', $definition->getArgument(5));
 
         $this->assertTrue($definition->hasTag('proxy'));
         $this->assertSame([